+++ src/pages/auth/Login.jsx
@@ -1,4 +1,3 @@
-// src/pages/Auth/Login.jsx
 import AuthForm from "components/auth/AuthForm";
 
 export default function Login() {
